proyectos_python/limpieza_datos/requirements.txt
snowflake-connector-python[pandas]==3.1.0
openpyxl==3.1.2
pyinstaller==5.13.0